Change vvsort

+++ tests/test_linalg.py
@@ -5,23 +5,7 @@ import dpnp as inp
 import numpy
 
 
-def vvsort(val, vec, size, xp):
-    # for i in range(size):
-    # imax = i
-    # for j in range(i + 1, size):
-    #     if xp.abs(val[imax]) < xp.abs(val[j]):
-    #         imax = j
-
-    # temp = val[i]
-    # val[i] = val[imax]
-    # val[imax] = temp
-
-    # for k in range(size):
-    #     temp = vec[k, i]
-    #     vec[k, i] = vec[k, imax]
-    #     vec[k, imax] = temp
-    
-    # The function was changed because 
+def vvsort(val, vec, size, xp): 
     for i in range(size):
         imax = i
         for j in range(i + 1, size):
@@ -33,14 +17,14 @@ def vvsort(val, vec, size, xp):
         unravel_i = numpy.unravel_index(i, val.shape)
         unravel_imax = numpy.unravel_index(imax, val.shape)
 
-        temp = xp.empty(tuple())
-        temp[()] = val[unravel_i] # make a copy
+        temp = xp.empty(tuple(), dtype=vec.dtype)
+        temp[()] = val[unravel_i]  # make a copy
         val[unravel_i] = val[unravel_imax]
         val[unravel_imax] = temp
         
         for k in range(size):
-            temp = xp.empty(tuple())
-            temp[()] = vec[k, i] # make a copy
+            temp = xp.empty(tuple(), dtype=val.dtype)
+            temp[()] = vec[k, i]  # make a copy
             vec[k, i] = vec[k, imax]
             vec[k, imax] = temp
